refactor(factions): split mobile directory into pure view + fetch wrapper (#743)
Extract FactionsDirectoryView, a pure prop-driven skin (factions, status,
invitations, loading, error, unaffiliated, onVisit) holding all the JSX — stripe
bar, invitation letters panel, unaffiliated banner and the FactionSelectCard
stack. DefaultFactionsDirectory stays the thin container: it keeps the useEffect
fetch of /factions + /factions/status + /factions/invitations and hands the
results down. Visual no-op — mirrors how DefaultPlayers is the pure skin behind
its container, so the surface is testable over controlled rows.
